external/boringssl: update #define guards for x86_64-gcc.c.
OS X builds with NO_ASM and was getting both generic.c and x86_64-gcc.c.
This change updates the latter so that it's excluded in NO_ASM builds.

This is a reland of 53b609c9, which got lost in the last BoringSSL sync
because I forgot to send it upstream.

Change-Id: I82462e5cd1d24fa96176c89d77cafc1f7ed0a0fd