external/boringssl: remove BORINGSSL_201509 support.
The BORINGSSL_201509 define was used to make updating BoringSSL in
external/boringssl less painful. It allowed code to compile with either
the old BoringSSL (which didn't define BORINGSSL_201509) or with the new
(which does).

Now that the new version has landed, this change removes that define. It
must be landed after the changes elsewhere in Android that remove
references to this define.

Change-Id: I19e661419f830459d015bf14e7905af2ec41b735